python-service/requirements.txt
kafka-python
clickhouse-connect
uuid7
